requirements-lock.txt
altgraph==0.17.4
anitopy==2.1.1
# Editable Git install with no remote (anivault==0.1.0)
-e f:\python_projects\anivault
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bandit==1.8.6
cattrs==25.2.0
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coverage==7.10.6
cryptography==46.0.0
detect-secrets==1.5.0
distlib==0.4.0
dnspython==2.8.0
email-validator==2.3.0
filelock==3.19.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hypothesis==6.140.2
identify==2.6.14
idna==3.10
iniconfig==2.1.0
Levenshtein==0.27.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
mypy==1.18.1
mypy_extensions==1.1.0
nodeenv==1.9.1
orjson==3.11.3
packaging==25.0
parse==1.20.2
pathspec==0.12.1
pbr==7.0.1
pefile==2023.2.7
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
psutil==7.1.0
py-cpuinfo==9.0.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyright==1.1.405
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-benchmark==5.1.0
pytest-cov==7.0.0
pytest-httpx==0.35.0
pytest-mock==3.15.1
pytest-qt==4.5.0
pytest_httpserver==1.1.3
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.14.1
readchar==4.2.1
requests==2.32.5
requests-cache==1.2.1
rich==14.1.0
ruff==0.13.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
specify-cli @ git+https://github.com/github/spec-kit.git@b18ef208cb8ba7b1eba04fe5dfcf3c24fda84a3e
SQLAlchemy==2.0.43
stevedore==5.5.0
tmdbsimple==2.9.1
tmdbv3api==1.9.0
toml==0.10.2
tomli_w==1.2.0
truststore==0.10.4
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
url-normalize==2.2.1
urllib3==2.5.0
uv==0.8.17
virtualenv==20.34.0
wcwidth==0.2.14
Werkzeug==3.1.3
